+++ CHANGELOG.md
@@ -7,9 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [2.8.0] - 2018-04-16
+
 ### Added
 
-  - `Timber.Integrations.ErrorLogger` allows you to include a new OTP `:error\_logger` handler to better maintain the structure of errors and stacktraces. It also collapses many cases of multi-line logs into a single line.
+  - `Timber.Integrations.ErrorLogger` allows you to include a new OTP
+    `:error\_logger` handler to better maintain the structure of errors and
+    stacktraces. It also collapses many cases of multi-line logs into a single
+    line.
 
 ## [2.7.0] - 2018-03-22
 
@@ -101,8 +106,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     `Timber.Integrations.PhoenixInstrumenter`. This will suppress log lines
     from being written for any controller/action pair.
 
-[Unreleased]: https://github.com/timberio/timber-elixir/compare/v2.7.0...HEAD
-[2.6.1]: https://github.com/timberio/timber-elixir/compare/v2.6.1...v2.7.0
+[Unreleased]: https://github.com/timberio/timber-elixir/compare/v2.8.0...HEAD
+[2.8.0]: https://github.com/timberio/timber-elixir/compare/v2.7.0...v2.8.0
+[2.7.0]: https://github.com/timberio/timber-elixir/compare/v2.6.1...v2.7.0
 [2.6.1]: https://github.com/timberio/timber-elixir/compare/v2.6.0...v2.6.1
 [2.6.0]: https://github.com/timberio/timber-elixir/compare/v2.5.6...v2.6.0
 [2.5.6]: https://github.com/timberio/timber-elixir/compare/v2.5.5...v2.5.6